+++ content-archive.php
@@ -19,7 +19,7 @@
 
 			if(is_home()):
                 $thumb_id = get_post_thumbnail_id();
-                $thumb_url_array = wp_get_attachment_image_src($thumb_id, 'blog-featured-image', true);
+                $thumb_url_array = wp_get_attachment_image_src($thumb_id, 'blog_featured_image', true);
                 $thumb_url = $thumb_url_array[0];
 
                 if(empty($thumb_url) || $thumb_url == SITEURL . '/wp-includes/images/media/default.png')  {