+++ functions.php
@@ -31,6 +31,7 @@
 		'add-image-to-post.php',              // Add featured image to post
 		'sherpa-video-embed-shortcode.php',   // Add shortcode 
 		'add-tags-to-single.php',             // Add tags to the content on single posts
+		'get-stars-for-rating.php',           // Get Stars for ratings on testimonials
 	);
 	
 	$prefix = 'inc/';