+++ inc/required-plugins.php
@@ -33,6 +33,11 @@ function sherpa_register_required_plugins() {
 	 * If the source is NOT from the .org repo, then source is also required.
 	 */
 	$plugins = array(
+    	array(
+			'name'      => 'Advanced Custom Fields',
+			'slug'      => 'advanced-custom-fields',
+			'required'  => false,
+    	),
 		array(
 			'name'      => 'Akismet',
 			'slug'      => 'akismet',
@@ -86,9 +91,9 @@ function sherpa_register_required_plugins() {
 	 * Only uncomment the strings in the config array if you want to customize the strings.
 	 */
 	$config = array(
-		'id'           => 'sherpa',                 // Unique ID for hashing notices for multiple instances of TGMPA.
+		'id'           => 'sherpa',                // Unique ID for hashing notices for multiple instances of TGMPA.
 		'default_path' => '',                      // Default absolute path to bundled plugins.
-		'menu'         => 'tgmpa-install-plugins', // Menu slug.
+		'menu'         => 'sherpa-install-plugins',// Menu slug.
 		'parent_slug'  => 'themes.php',            // Parent menu slug.
 		'capability'   => 'edit_theme_options',    // Capability needed to view plugin install page, should be a capability associated with the parent menu used.
 		'has_notices'  => true,                    // Show admin notices or not.